backend/pyproject.toml
[project]
name = "reach-backend"
version = "0.1.0"
description = "FastAPI backend for the reach assignment"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.18.4",
    "asyncpg>=0.30.0",
    "email-validator>=2.3.0",
    "fastapi>=0.111.0",
    "psycopg2-binary>=2.9.12",
    "pwdlib[argon2]>=0.3.0",
    "pydantic-settings>=2.7.0",
    "pyjwt>=2.12.1",
    "python-multipart>=0.0.27",
    "sqlalchemy[asyncio]>=2.0.0",
    "uvicorn[standard]>=0.30.0",
]

[dependency-groups]
dev = [
    "httpx>=0.27.0",
    "pytest>=8.2.0",
]

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
